refactor(parseAddress): move parseAddress to DefaultPolicy
