+++ lib/amqp_client.js
@@ -93,7 +93,7 @@ AMQPClient.prototype.connect = function(url) {
 
     debug('connecting to: ' + url);
     self._reconnect = self.connect.bind(self, url);
-    var address = u.parseAddress(url);
+    var address = self.policy.parseAddress(url);
     self._defaultQueue = address.path.substr(1);
     self.policy.connect.options.hostname = address.host;
     var sasl = address.user ? new Sasl() : null;
