+++ lib/connection.js
@@ -522,7 +522,7 @@ Connection.prototype._connect = function(address, sasl) {
     self.client = tls.connect(sslOptions);
     debug('Connecting to ' + self.address.host + ':' + self.address.port + ' via TLS');
   } else {
-    self.client = net.connect({port: self.address.port, host: self.address.host});
+    self.client = net.connect({ port: self.address.port, host: self.address.host });
     debug('Connecting to ' + self.address.host + ':' + self.address.port + ' via straight-up sockets');
   }
   self.client.on(isSSL ? 'secureConnect' : 'connect', function() {
