+++ lib/policies/default_policy.js
@@ -1,6 +1,7 @@
 'use strict';
 
-var constants = require('../constants'),
+var url = require('url'),
+    constants = require('../constants'),
     putils = require('./policy_utilities');
 
 function containerName() {
@@ -38,6 +39,7 @@ module.exports = {
     strategy: 'fibonacci', // || 'exponential'
     forever: true
   },
+
   connect: {
     options: {
       containerId: containerName(),
@@ -58,6 +60,7 @@ module.exports = {
       }
     }
   },
+
   session: {
     options: {
       nextOutgoingId: constants.session.defaultOutgoingId,
@@ -68,6 +71,7 @@ module.exports = {
     windowQuantum: constants.session.defaultIncomingWindow,
     enableSessionFlowControl: true
   },
+
   senderLink: {
     attach: {
       name: linkName('sender'),
@@ -80,6 +84,7 @@ module.exports = {
     encoder: function(body) { return body; },
     reattach: null
   },
+
   receiverLink: {
     attach: {
       name: linkName('receiver'),
@@ -92,5 +97,39 @@ module.exports = {
     creditQuantum: 100,
     decoder: function(body) { return body; },
     reattach: null
+  },
+
+  /**
+   * Parses an address for use when connecting to an AMQP 1.0 broker
+   */
+  parseAddress: function(address) {
+    var parsedAddress = url.parse(address);
+    var result = {
+      host: parsedAddress.hostname,
+      path: parsedAddress.path || '/',
+      protocol: parsedAddress.protocol.slice(0, -1).toLowerCase() || 'amqp',
+    };
+
+    if (result.protocol !== 'amqp' && result.protocol !== 'amqps')
+      throw new Error('invalid protocol: ' + result.protocol);
+
+    if (!!parsedAddress.port) {
+      result.port = parseInt(parsedAddress.port);
+    } else {
+      switch (result.protocol.toLowerCase()) {
+        case 'amqp': result.port = constants.defaultPort; break;
+        case 'amqps': result.port = constants.defaultTlsPort; break;
+      }
+    }
+
+    result.rootUri = parsedAddress.protocol + '//';
+    if (!!parsedAddress.auth) {
+      result.rootUri += parsedAddress.auth + '@';
+      result.user = parsedAddress.auth.split(':')[0];
+      result.pass = parsedAddress.auth.split(':')[1] || null;
+    }
+
+    result.rootUri += result.host + ':' + result.port;
+    return result;
   }
 };
