+++ lib/utilities.js
@@ -2,13 +2,9 @@
 
 var _ = require('lodash'),
     Int64 = require('node-int64'),
-
-    constants = require('./constants'),
-    errors = require('./errors'),
-
     uuid = require('uuid'),
+    errors = require('./errors'),
     utilities = module.exports = {};
-
 
 /**
  * Encodes given value into node-amqp-encoder format.
@@ -91,62 +87,6 @@ function bufferEquals(lhs, rhs, offset1, offset2, size) {
 }
 
 utilities.bufferEquals = bufferEquals;
-
-// Constants
-var addressRegex = new RegExp('^(amqps?)://([^:/]+)(?::([0-9]+))?(/.*)?$');
-var addressWithCredentialsRegex = new RegExp('^(amqps?)://([^:]+):([^@]+)@([^:/]+)(?::([0-9]+))?(/.*)?$');
-
-function getPort(port, protocol) {
-  if (port) {
-    var asFloat = parseFloat(port);
-    if (!isNaN(asFloat) && isFinite(port) && (port % 1 === 0)) {
-      return asFloat;
-    } else {
-      throw new Error('Invalid port: ' + port);
-    }
-  } else {
-    switch (protocol) {
-      case 'amqp':
-        return constants.defaultPort;
-      case 'amqps':
-        return constants.defaultTlsPort;
-      default:
-        throw new Error('Unknown Protocol ' + protocol);
-    }
-  }
-}
-
-function parseAddress(address) {
-  var results = addressWithCredentialsRegex.exec(address);
-  if (results) {
-    results = {
-      protocol: results[1],
-      user: decodeURIComponent(results[2]),
-      pass: decodeURIComponent(results[3]),
-      host: results[4],
-      port: getPort(results[5], results[1]),
-      path: results[6] || '/'
-    };
-    results.rootUri = results.protocol + '://' + results.user + ':' + results.pass + '@' + results.host + ':' + results.port;
-  } else {
-    results = addressRegex.exec(address);
-    if (results) {
-      results = {
-        protocol: results[1],
-        host: results[2],
-        port: getPort(results[3], results[1]),
-        path: results[4] || '/'
-      };
-      results.rootUri = results.protocol + '://' + results.host + ':' + results.port;
-    }
-  }
-
-  if (results) return results;
-
-  throw new Error('Failed to parse ' + address);
-}
-
-utilities.parseAddress = parseAddress;
 
 // @todo: this "parsing" should be far more rigorous
 utilities.parseLinkAddress = function(address, policy) {
